apps/worker/requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.0
httpx==0.27.0
pytest==8.0.0
